YEL-139 [test] Factory 테스트 코드 작성
