+++ src/test/java/com/yello/server/global/common/factory/TimeFactoryTest.java
@@ -2,26 +2,41 @@ package com.yello.server.global.common.factory;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.concurrent.ThreadLocalRandom;
+import java.time.Duration;
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
-@DisplayName("WeightedRandomFactory 에서")
+@DisplayName("TimeFactory 에서")
 @DisplayNameGeneration(ReplaceUnderscores.class)
-public class WeightedRandomFactoryTest {
+public class TimeFactoryTest {
 
     @Test
-    void 가중치_랜덤값_생성에_성공합니다() {
+    void 시간을_문자열_형식으로_포맷팅에_성공합니다() {
         // given
-        final ThreadLocalRandom random = ThreadLocalRandom.current();
+        final LocalDateTime localDateTime = LocalDateTime.of(2023, 1, 1, 14, 0, 0);
+
+        final LocalDateTime currentDateTime = LocalDateTime.now();
+        Duration duration = Duration.between(localDateTime, currentDateTime);
+        long seconds = duration.getSeconds();
+        String expect = "";
+
+        if (seconds < 60) {
+            expect = seconds + "초 전";
+        } else if ((seconds /= 60) < 60) {
+            expect = seconds + "분 전";
+        } else if ((seconds /= 60) < 24) {
+            expect = (seconds) + "시간 전";
+        }
+
+        expect = (seconds / 24) + "일 전";
 
         // when
-        Integer result = WeightedRandomFactory.randomPoint();
+        String result = TimeFactory.toFormattedString(localDateTime);
 
         // then
-        assertThat(result).isBetween(5, 25);
-        assertThat(random.nextDouble()).isBetween(0.0, 1.0);
+        assertThat(result).isEqualTo(expect);
     }
 }
